- migrate endpoint to report area
