+++ source/backend/api/Areas/Reports/Controllers/AcquisitionController.cs
@@ -1,14 +1,19 @@
+using System;
 using System.Linq;
 using System.Security.Claims;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.Extensions.Logging;
+using Pims.Api.Areas.Acquisition.Models.Search;
 using Pims.Api.Areas.Reports.Models.Agreement;
 using Pims.Api.Helpers.Constants;
 using Pims.Api.Helpers.Exceptions;
 using Pims.Api.Helpers.Extensions;
 using Pims.Api.Helpers.Reporting;
+using Pims.Api.Models;
 using Pims.Api.Policies;
 using Pims.Api.Services;
+using Pims.Core.Extensions;
 using Pims.Dal.Entities.Models;
 using Pims.Dal.Security;
 using Swashbuckle.AspNetCore.Annotations;
@@ -29,6 +34,7 @@ namespace Pims.Api.Areas.Reports.Controllers
         #region Variables
         private readonly IAcquisitionFileService _acquisitionFileService;
         private readonly ClaimsPrincipal _user;
+        private readonly ILogger _logger;
         #endregion
 
         #region Constructors
@@ -37,10 +43,13 @@ namespace Pims.Api.Areas.Reports.Controllers
         /// Creates a new instance of a AcquisitionController class, initializes it with the specified arguments.
         /// </summary>
         /// <param name="acquisitionFileService"></param>
-        public AcquisitionController(IAcquisitionFileService acquisitionFileService, ClaimsPrincipal user)
+        /// <param name="user"></param>
+        /// <param name="logger"></param>
+        public AcquisitionController(IAcquisitionFileService acquisitionFileService, ClaimsPrincipal user, ILogger<AcquisitionController> logger)
         {
             _acquisitionFileService = acquisitionFileService;
             _user = user;
+            _logger = logger;
         }
         #endregion
 
@@ -73,6 +82,40 @@ namespace Pims.Api.Areas.Reports.Controllers
             var reportAgreements = agreements.Select(agreement => new AgreementReportModel(agreement, _user));
 
             return ReportHelper.GenerateExcel(reportAgreements, "Agreement Export");
+        }
+
+        [HttpGet]
+        [HasPermission(Permissions.AcquisitionFileView)]
+        [Produces(ContentTypes.CONTENTTYPEEXCELX)]
+        [ProducesResponseType(200)]
+        [ProducesResponseType(typeof(ErrorResponseModel), 409)]
+        [SwaggerOperation(Tags = new[] { "acquisitionfile", "report" })]
+        public IActionResult ExportAcquisitionFiles([FromQuery] AcquisitionFilterModel filter)
+        {
+            _logger.LogInformation(
+                "Request received by Controller: {Controller}, Action: {ControllerAction}, User: {User}, DateTime: {DateTime}",
+                nameof(AcquisitionController),
+                nameof(ExportAcquisitionFiles),
+                _user.GetUsername(),
+                DateTime.Now);
+
+            _logger.LogInformation("Dispatching to service: {Service}", _acquisitionFileService.GetType());
+
+            filter.ThrowBadRequestIfNull($"The request must include a filter.");
+            if (!filter.IsValid())
+            {
+                throw new BadRequestException("Acquisition files filter must contain valid values.");
+            }
+
+            var acceptHeader = (string)Request.Headers["Accept"];
+            if (acceptHeader != ContentTypes.CONTENTTYPEEXCEL && acceptHeader != ContentTypes.CONTENTTYPEEXCELX)
+            {
+                throw new BadRequestException($"Invalid HTTP request header 'Accept:{acceptHeader}'.");
+            }
+
+            var acquisitionFileData = _acquisitionFileService.GetAcquisitionFileExport((AcquisitionFilter)filter);
+
+            return ReportHelper.GenerateExcel(acquisitionFileData, "PIMS_Acquisition_Files");
         }
         #endregion
     }
